requirements.txt
discord.py
flask
requests